# Speed up Relate for PreparedGeometry